Refactor models to src directory and refactor faq validation to use class validator and new validateModel method
