+++ package.json
@@ -31,6 +31,8 @@
         "aws-sdk": "^2.724.0",
         "bcryptjs": "^2.4.3",
         "bluebird": "^3.7.2",
+        "class-transformer": "^0.3.2",
+        "class-validator": "^0.13.1",
         "colors": "^1.4.0",
         "cookie-parser": "^1.4.5",
         "cors": "^2.8.5",
