--- src/database/models/faqs.model.ts
@@ -1,17 +1,0 @@
-export interface Faq {
-    faq_id: number;
-    listing_id: string;
-    question: string;
-    answer: string;
-}
-
-export interface CreateFaqSchema {
-    listing_id: string;
-    question: string;
-    answer: string;
-}
-
-export interface UpdateFaqSchema {
-    question: string;
-    answer: string;
-}
