+++ src/database/repositories/categories.repository.ts
@@ -1,5 +1,5 @@
 import { BaseRepository } from './base.repository';
-import { Category } from '../models';
+import { Category } from '../../models';
 
 export class CategoriesRepository extends BaseRepository {
     list(): Promise<Category[]> {
