+++ src/models/hashtags.model.ts
