+++ src/models/jobs.model.ts
