+++ src/models/listings.model.ts
